+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF0T8PJX7TSBHHMABCZ09S19"
+        "01KF0V8C5DBFDQY3HWYW8PFR2R"
     )
     r"""The unique identifier of the data source"""