+++ packages/orq-rc/RELEASES.md
@@ -4699,3 +4699,13 @@ Based on:
 - [python v4.2.0-rc.32] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.32 - packages/orq-rc
+
+## 2026-01-15 12:51:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.685.3 (2.795.8) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.33] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.33] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.33 - packages/orq-rc